Moved `getRecentChanges` to its own ViewModel.
